pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "teleGemmaBot"
version = "0.1.0"
description = "Telegram bot bridge to Google Gemma API"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0,<22.0",
    "google-generativeai>=0.8.0,<1.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=5.0,<6.0",
    "ruff>=0.5,<1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
